Giant refactor
Removed the Flask Server and refactored to use
a Node Express Server with Socket Io
I spent all day trying to figure out why the
front end wouldn't accept live updates, only to
reimplement everything in TS/JS and have it work
as expected, not sure what that is about.

This is still in progress but on page load, the client
connects to the socket server and is sent the list of
recently scanned items.

When an item is added, the main process of Trashscan
sends an API request to the update endpoint on the
express server and then socket io, fires a message
to the connected client. This cuts down on the amount
of calls needed, and now we get live updates exactly
when they happen with out the need to poll. This also
achieves persistence for us since we no longer need
to duplicate our efforts.