Updated The GUI
Added drop shadows to the first card, or the last scanned item.
Modified the Socket Provider to refresh the connection every
four minutes. Moved the refresh button to the left so it is more
visible. These changes are ready to be tested and if work fine
this branch can be merged into main.